arch. layout 작성
