+++ ditto-module-user/src/main/java/org/example/service/MyService.java
@@ -13,6 +13,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -64,5 +66,9 @@ public class MyService {
         } else {
             throw new InvalidAccessException();
         }
+    }
+
+    public List<Map<String, Object>> findMembers() {
+        return memberMapper.findMembers();
     }
 }
