+++ ditto-module-user/src/main/resources/templates/layouts/layout-sample.html
@@ -1,0 +1,98 @@
+<!DOCTYPE html>
+<html lang="ko"
+      xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+
+<head>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" >
+    <link rel='stylesheet' href='https://fonts.googleapis.com/css?family=Roboto'>
+    <script src='https://kit.fontawesome.com/a076d05399.js' crossorigin='anonymous'></script>
+    <style>
+        body {
+          font-family: Arial, Helvetica, sans-serif;
+          margin: 0;
+          background-color: #f4f4f4;
+        }
+        .body-container {
+          margin: 0;
+          padding: 0;
+        }
+        .main-container {
+          margin: auto;
+          margin-left: 160px;
+          max-width: 100%;
+        }
+        .side-container {
+          height: 100%;
+          width: 160px;
+          position: fixed;
+          z-index: 1;
+          background-color: #212529;
+          overflow-x: hidden;
+          padding-top: 10px;
+          top: 0;
+          left: 0;
+          margin: 0;
+          padding: 0;
+        }
+        .content-container {
+          padding: 20px;
+          width: 100%;
+        }
+        .main-title {
+          font-size: 1.4rem;
+          font-weight: 600;
+        }
+        .portlet {
+          padding: 10px 20px;
+          width: 100%;
+          border-radius: 5px;
+          background-color: #fff;
+        }
+        .vertical-menu {
+          width: 100%; /* Set a width if you like */
+        }
+        .vertical-menu a {
+          color: white;
+          display: block; /* Make the links appear below each other */
+          padding: 10px;
+          text-decoration: none; /* Remove underline from links */
+        }
+        .vertical-menu a:hover {
+          color: black;
+          background-color: #ccc;/* Dark grey background on mouse-over */
+        }
+        .tb-success {
+          border-collapse: collapse;
+          width: 100%;
+        }
+        .tb-success td, .tb-success th {
+          border: 1px solid #ddd;
+          padding: 4px;
+        }
+        .tb-success tr:nth-child(even){background-color: #f2f2f2;}
+        .tb-success tr:hover {background-color: #ddd;}
+        .tb-success th {
+          padding-top: 6px;
+          padding-bottom: 6px;
+          text-align: left;
+          background-color: #04AA6D;
+          color: white;
+        }
+    </style>
+</head>
+<body>
+<div class="body-container">
+
+    <aside th:replace="layouts/sample-aside :: aside"></aside>
+
+    <div class="main-container">
+        <div class="content-container">
+            <th:block layout:fragment="content"></th:block>
+            <footer th:replace="layouts/sample-footer :: footer"></footer>
+        </div>
+    </div>
+</div>
+</body>
+</html>
