+++ ditto-module-user/src/main/resources/templates/layouts/sample-aside.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<th:block th:fragment="aside">
+    <aside class="side-container">
+        <div class="vertical-menu ">
+            <button type="button" class="btn btn-success w-100 text-start"><span class="fa fa-home"></span> Home</button>
+            <hr>
+            <a href="/members"><i class="fas fa-arrow-alt-circle-right"></i><span>&nbsp; Members</span></a>
+        </div>
+    </aside>
+</th:block>
+</html>
