+++ ditto-module-user/src/main/resources/templates/layouts/sample-footer.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+
+<th:block th:fragment="footer">
+    <!-- Footer -->
+    <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
+        <p class="col-md-4 mb-0 text-muted">&copy; 2022 onda2me, </p>
+        <ul class="nav col-md-4 justify-content-end">
+            <li class="nav-item"><a href="/" class="nav-link px-2 text-muted" target="_blank">Home</a></li>
+            <li class="nav-item"><a href="#" class="nav-link px-2 text-muted" target="_blank">About</a></li>
+        </ul>
+    </footer>
+    <!-- Footer -->
+</th:block>
+</html>
